+++ test/correctness/PySys_internal_121/run.py
@@ -27,6 +27,7 @@ waitForSignal
 logFileContentsDefaultExcludes
 getExprFromFile
 pysys
+threadPoolMaxWorkers
 		""".strip().split('\n')
 
 		def filter_member(m):